+++ packages/tab-renderer/src/react/TabLine.tsx
@@ -1,20 +1,23 @@
 import type { TabLineProps } from "./types";
 import { TabChord } from "./TabChord";
+import { TabDecoration } from "./TabDecoration";
 import { TabLyric } from "./TabLyric";
 
 export function TabLine({ line, className }: TabLineProps) {
-	return (
-		<div className={className} style={{ whiteSpace: "pre" }}>
-			{line.tokens.map((token, tokenIndex) => {
-				switch (token.kind) {
-					case "ChordToken":
-						return <TabChord key={tokenIndex} token={token} />;
-					case "LyricToken":
-						return <TabLyric key={tokenIndex} token={token} />;
-					default:
-						return <span key={tokenIndex}>{token.text}</span>;
-				}
-			})}
-		</div>
-	);
+  return (
+    <div className={className} style={{ whiteSpace: "pre" }}>
+      {line.tokens.map((token, tokenIndex) => {
+        switch (token.kind) {
+          case "ChordToken":
+            return <TabChord key={tokenIndex} token={token} />;
+          case "LyricToken":
+            return <TabLyric key={tokenIndex} token={token} />;
+          case "DecorationToken":
+            return <TabDecoration key={tokenIndex} token={token} />;
+          default:
+            return <span key={tokenIndex}>{token.text}</span>;
+        }
+      })}
+    </div>
+  );
 }