+++ packages/tab-renderer/src/react/index.ts
@@ -4,6 +4,7 @@ export { TabSection } from "./TabSection";
 export { TabLine } from "./TabLine";
 export { TabChord } from "./TabChord";
 export { TabLyric } from "./TabLyric";
+export { TabDecoration } from "./TabDecoration";
 export {
   DEFAULT_TAB_STYLE,
   clampContentMarginRightPx,